+++ java/com/hmdzl/spspd/items/misc/CopyBall.java
@@ -32,7 +32,7 @@ import com.hmdzl.spspd.effects.Splash;
 import com.hmdzl.spspd.items.Item;
 import com.hmdzl.spspd.items.bombs.DungeonBomb;
 import com.hmdzl.spspd.items.weapon.missiles.MissileWeapon;
-import com.hmdzl.spspd.messages.Messages;
+import com.hmdzl.spspd.messages.Messages;import com.hmdzl.spspd.ResultDescriptions;
 import com.hmdzl.spspd.scenes.CellSelector;
 import com.hmdzl.spspd.scenes.GameScene;
 import com.hmdzl.spspd.sprites.ItemSpriteSheet;
@@ -95,7 +95,7 @@ public class CopyBall extends Item {
 
 	@Override
 	public String status() {
-		return Messages.format("%d", (int)charge/10);
+		return Messages.format("%d", charge /10);
 	}
 	
 	@Override
