+++ java/com/hmdzl/spspd/items/misc/DiceTower.java
@@ -20,29 +20,16 @@ package com.hmdzl.spspd.items.misc;
 import java.util.ArrayList;
 
 import com.hmdzl.spspd.Dungeon;
-import com.hmdzl.spspd.actors.Actor;
-import com.hmdzl.spspd.actors.Char;
-import com.hmdzl.spspd.actors.buffs.ArmorBreak;
-import com.hmdzl.spspd.actors.buffs.AttackDown;
-import com.hmdzl.spspd.actors.buffs.Bleeding;
-import com.hmdzl.spspd.actors.buffs.Buff;
-import com.hmdzl.spspd.actors.buffs.Cripple;
-import com.hmdzl.spspd.actors.buffs.Poison;
-import com.hmdzl.spspd.actors.buffs.Weakness;
 import com.hmdzl.spspd.actors.hero.Hero;
-import com.hmdzl.spspd.items.Generator;
 import com.hmdzl.spspd.items.Item;
 
-import com.hmdzl.spspd.levels.Level;
-import com.hmdzl.spspd.messages.Messages;
+import com.hmdzl.spspd.messages.Messages;import com.hmdzl.spspd.ResultDescriptions;
 import com.hmdzl.spspd.scenes.GameScene;
 import com.hmdzl.spspd.sprites.ItemSpriteSheet;
 import com.hmdzl.spspd.utils.GLog;
 import com.hmdzl.spspd.windows.WndItem;
 import com.watabou.utils.Bundle;
 import com.watabou.utils.Random;
-
-import static com.hmdzl.spspd.ui.ActionIndicator.action;
 
 public class DiceTower extends Item {
 
@@ -141,7 +128,7 @@ public class DiceTower extends Item {
 	
 	 @Override
 	 public String status() {
-			 return Messages.format("%d", (int)charge/60);
+			 return Messages.format("%d", charge /60);
 	 }		
 
 }
