+++ java/com/hmdzl/spspd/items/misc/GunOfSoldier.java
@@ -26,7 +26,7 @@ import com.hmdzl.spspd.actors.hero.Hero;
 import com.hmdzl.spspd.effects.Splash;
 import com.hmdzl.spspd.items.Item;
 import com.hmdzl.spspd.items.weapon.missiles.MissileWeapon;
-import com.hmdzl.spspd.messages.Messages;
+import com.hmdzl.spspd.messages.Messages;import com.hmdzl.spspd.ResultDescriptions;
 import com.hmdzl.spspd.scenes.CellSelector;
 import com.hmdzl.spspd.scenes.GameScene;
 import com.hmdzl.spspd.sprites.ItemSpriteSheet;
@@ -90,7 +90,7 @@ public class GunOfSoldier extends Item {
 
 	@Override
 	public String status() {
-		return Messages.format("%d", (int)charge/75);
+		return Messages.format("%d", charge /75);
 	}
 	
 	@Override
